extract reusable recurrence fields
